Add internal flag pole provider
