+++ tests/sentry/flags/test_audit_log_presenter.py
@@ -1,13 +1,12 @@
 from django.urls import reverse
 
 from sentry.flags.models import ACTION_MAP, CREATED_BY_TYPE_MAP, FlagAuditLogModel
-from sentry.runner.commands.presenters.audit_log_presenter import AuditLogPresenter, AuditLogRequest
+from sentry.runner.commands.presenters.audit_log_presenter import AuditLogPresenter
 from sentry.testutils.cases import APITestCase
-from sentry.utils.security.orgauthtoken_token import hash_token
 
 
 def test_audit_log_item_generation():
-    presenter = AuditLogPresenter("", None)
+    presenter = AuditLogPresenter("")
     presenter.set("a", True)
     presenter.unset("b")
     presenter.update("c", True, False)
@@ -41,33 +40,13 @@ class AuditLogPresenterFunctionalTestCase(APITestCase):
         self.url = reverse(self.endpoint, args=(self.organization.slug, "flag-pole"))
 
     def test_post(self):
-        def mock_request(request: AuditLogRequest):
-            self.client.post(
-                request["url"],
-                data={"data": request["data"]},
-                headers=request["headers"],
-            )
-
-        token = "sntrys_abc123_xyz"
-        self.create_org_auth_token(
-            name="Test Token 1",
-            token_hashed=hash_token(token),
-            organization_id=self.organization.id,
-            token_last_characters="xyz",
-            scope_list=["org:ci"],
-            date_last_used=None,
-        )
-
-        self.client.credentials(HTTP_AUTHORIZATION=f"Bearer {token}")
-
         with self.options(
             {
-                "flags:options-audit-log-webhook-url": self.url,
-                "flags:options-audit-log-token": token,
-                "flags:options-audit-log-is-disabled": False,
+                "flags:options-audit-log-is-enabled": True,
+                "flags:options-audit-log-organization-id": self.organization.id,
             }
         ):
-            presenter = AuditLogPresenter("", request_fn=mock_request)
+            presenter = AuditLogPresenter("")
             presenter.set("a", True)
             presenter.flush()
 
@@ -77,6 +56,6 @@ class AuditLogPresenterFunctionalTestCase(APITestCase):
         assert flag.action == ACTION_MAP["created"]
         assert flag.flag == "a"
         assert flag.created_by == "internal"
-        assert flag.created_by_type == CREATED_BY_TYPE_MAP["email"]
+        assert flag.created_by_type == CREATED_BY_TYPE_MAP["name"]
         assert flag.organization_id == self.organization.id
         assert flag.tags == {"value": True}
